Backend/requirements.txt
# Core ML - Chronos
#git+https://github.com/amazon-science/chronos-forecasting.git
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Technical indicators
ta>=0.11.0

# API & Web
flask>=3.0.0
python-binance>=1.0.19
python-dotenv>=1.0.0
requests>=2.31.0

# Utilities
joblib>=1.3.0
tqdm>=4.66.0

# Logging
loguru>=0.7.0

# Scheduling
schedule>=1.2.0

# Optional visualization
matplotlib>=3.8.0
plotly>=5.17.0

# XGBoost for ML models
xgboost>=2.0.0